+++ fusen-internal-common/src/lib.rs
@@ -5,6 +5,7 @@
 pub mod protocol;
 #[allow(missing_docs)]
 pub mod resource;
+#[cfg(feature = "uuid")]
 #[allow(missing_docs)]
 pub mod utils;
 
@@ -14,4 +15,5 @@ pub type BoxFuture<T> = std::pin::Pin<Box<dyn std::future::Future<Output = T> + 
 pub type BoxFutureV2<'a, T> = std::pin::Pin<Box<dyn std::future::Future<Output = T> + Send + 'a>>;
 
 /// JSON types re-exported for generated code.
+#[cfg(feature = "serde-json")]
 pub use serde_json;