+++ fusen-internal-common/src/utils/uuid.rs
@@ -1,6 +1,5 @@
-use crate::utils::uuid;
 use ::uuid::Uuid;
 
 pub fn uuid() -> String {
-    uuid::Uuid::new_v4().to_string()
+    Uuid::new_v4().to_string()
 }